refactor: change the path of the loading component
