+++ src/components/loading/form/index.js
@@ -1,0 +1,7 @@
+import { SLoading } from '../styled'
+
+const Loading = () => {
+  return <SLoading />
+}
+
+export default Loading
