+++ src/components/loading/page/index.js
@@ -1,0 +1,11 @@
+import { SLoading, LoadingBox } from '../styled'
+
+const Loading = () => {
+  return (
+    <LoadingBox>
+      <SLoading />
+    </LoadingBox>
+  )
+}
+
+export default Loading
